Create riskIQ_hostpair_api
This script will search host pairs from a list of domains you paste into a csv file.  Domain names accepted are not case sensitive, and are searchable as: example.com. No need to put: 'https://www.'.

The error handling is basic with a view to skipping domains that aren't found. Improvements could be made by creating a function that checks the domain status code before running the API call.